Modify data types for optimization

+++ orders/models.py
@@ -21,7 +21,7 @@ class Order(models.Model):
     created = models.DateTimeField(auto_now_add=True)
     updated = models.DateTimeField(auto_now=True)
     paid = models.BooleanField(default=False)
-    final_price = models.IntegerField()
+    final_price = models.PositiveIntegerField()
 
     class Meta:
         ordering = ('-created',)
@@ -36,7 +36,7 @@ class OrderItem(models.Model):
     order = models.ForeignKey(Order, related_name='items', on_delete=models.CASCADE)
     product = models.ForeignKey(Product, related_name='order_items', on_delete=models.CASCADE)
     price = models.DecimalField(max_digits=10, decimal_places=2)
-    quantity = models.PositiveIntegerField(default=1)
+    quantity = models.PositiveSmallIntegerField(default=1)
 
     def __str__(self):
         return str(self.id)
